Fix bug in Quote model
I was curious why the rubric specified that quotes were never returned
with episode ids. I was trying to mimic this by removing it from the
json blob in the sql query, but it was really being taken out anyway
because the Quote model just throws out the episode id. So instead I
fixed the model and made the queries filter out the redundant id.